+++ src/components/radio.js
@@ -121,6 +121,10 @@
     }, []);
 
     useEffect(() => {
+      B.triggerEvent('onChange', value);
+    }, [value]);
+
+    useEffect(() => {
       if (mounted.current && loading) {
         B.triggerEvent('onLoad', loading);
       }