+++ src/components/select.js
@@ -86,7 +86,7 @@
           ...(orderBy ? { sort: { relation: sort } } : {}),
         },
         onCompleted(res) {
-          const hasResult = res && res.result && res.result.length > 0;
+          const hasResult = res && res.results && res.results.length > 0;
           if (hasResult) {
             B.triggerEvent('onSuccess', res.results);
           } else {
@@ -106,6 +106,10 @@
         mounted.current = false;
       };
     }, []);
+
+    useEffect(() => {
+      B.triggerEvent('onChange', currentValue);
+    });
 
     useEffect(() => {
       if (mounted.current && loading) {