+++ src/prefabs/card.js
@@ -68,6 +68,12 @@
             },
           },
         },
+        {
+          value: ['0rem', '0rem', '0rem', '0rem'],
+          label: 'Outer space',
+          key: 'outerSpacing',
+          type: 'SIZES',
+        },
       ],
       descendants: [
         {